feat(component): reuse rc-dock logic; allow zoom from click position
The modal needs to be accessed for its positions & dimensions during the transition.
Unfortunately, that kind of information is unavailable until after the modal mounts!
The modal mounts only after the state has changed and the render occurs, which means
a setState()-->callback-->setState() is required.

Using .setState() callback proved troublesome, though. It makes the whole component
render twice, so the modal appears & disappears in one click. This is unwanted behavior.

An alternative would be to imitate rc-dock by reading the DOM & applying styles
to the modal directly during every .setState() call. Although it goes against React principles
by attacking the DOM directly, it will do the job for now.